Add --split-configs
