+++ app/templates/Gruntfile.js
@@ -20,406 +20,34 @@ module.exports = function (grunt) {
         app: 'app',
         dist: 'dist'
     };
-
+    <% if (splitConfigs){ %>
+    config = require('load-grunt-configs')(grunt, config);
+    grunt.initConfig(config);
+    <% }else{ %>
     // Define the configuration for all the tasks
     grunt.initConfig({
-
-        // Project settings
-        config: config,
-
-        // Watches files for changes and runs tasks based on the changed files
-        watch: {
-            bower: {
-                files: ['bower.json'],
-                tasks: ['bowerInstall']
-            },<% if (coffee) { %>
-            coffee: {
-                files: ['<%%= config.app %>/scripts/{,*/}*.{coffee,litcoffee,coffee.md}'],
-                tasks: ['coffee:dist']
-            },
-            coffeeTest: {
-                files: ['test/spec/{,*/}*.{coffee,litcoffee,coffee.md}'],
-                tasks: ['coffee:test', 'test:watch']
-            },<% } else { %>
-            js: {
-                files: ['<%%= config.app %>/scripts/{,*/}*.js'],
-                tasks: ['jshint'],
-                options: {
-                    livereload: true
-                }
-            },
-            jstest: {
-                files: ['test/spec/{,*/}*.js'],
-                tasks: ['test:watch']
-            },<% } %>
-            gruntfile: {
-                files: ['Gruntfile.js']
-            },<% if (includeSass) { %>
-            sass: {
-                files: ['<%%= config.app %>/styles/{,*/}*.{scss,sass}'],
-                tasks: ['sass:server', 'autoprefixer']
-            },<% } %>
-            styles: {
-                files: ['<%%= config.app %>/styles/{,*/}*.css'],
-                tasks: ['newer:copy:styles', 'autoprefixer']
-            },
-            livereload: {
-                options: {
-                    livereload: '<%%= connect.options.livereload %>'
-                },
-                files: [
-                    '<%%= config.app %>/{,*/}*.html',
-                    '.tmp/styles/{,*/}*.css',<% if (coffee) { %>
-                    '.tmp/scripts/{,*/}*.js',<% } %>
-                    '<%%= config.app %>/images/{,*/}*'
-                ]
-            }
-        },
-
-        // The actual grunt server settings
-        connect: {
-            options: {
-                port: 9000,
-                open: true,
-                livereload: 35729,
-                // Change this to '0.0.0.0' to access the server from outside
-                hostname: 'localhost'
-            },
-            livereload: {
-                options: {
-                    middleware: function(connect) {
-                        return [
-                            connect.static('.tmp'),
-                            connect().use('/bower_components', connect.static('./bower_components')),
-                            connect.static(config.app)
-                        ];
-                    }
-                }
-            },
-            test: {
-                options: {
-                    open: false,
-                    port: 9001,
-                    middleware: function(connect) {
-                        return [
-                            connect.static('.tmp'),
-                            connect.static('test'),
-                            connect().use('/bower_components', connect.static('./bower_components')),
-                            connect.static(config.app)
-                        ];
-                    }
-                }
-            },
-            dist: {
-                options: {
-                    base: '<%%= config.dist %>',
-                    livereload: false
-                }
-            }
-        },
-
-        // Empties folders to start fresh
-        clean: {
-            dist: {
-                files: [{
-                    dot: true,
-                    src: [
-                        '.tmp',
-                        '<%%= config.dist %>/*',
-                        '!<%%= config.dist %>/.git*'
-                    ]
-                }]
-            },
-            server: '.tmp'
-        },
-
-        // Make sure code styles are up to par and there are no obvious mistakes
-        jshint: {
-            options: {
-                jshintrc: '.jshintrc',
-                reporter: require('jshint-stylish')
-            },
-            all: [
-                'Gruntfile.js',
-                '<%%= config.app %>/scripts/{,*/}*.js',
-                '!<%%= config.app %>/scripts/vendor/*',
-                'test/spec/{,*/}*.js'
-            ]
-        },<% if (testFramework === 'mocha') { %>
-
-        // Mocha testing framework configuration options
-        mocha: {
-            all: {
-                options: {
-                    run: true,
-                    urls: ['http://<%%= connect.test.options.hostname %>:<%%= connect.test.options.port %>/index.html']
-                }
-            }
-        },<% } else if (testFramework === 'jasmine') { %>
-
-        // Jasmine testing framework configuration options
-        jasmine: {
-            all: {
-                options: {
-                    specs: 'test/spec/{,*/}*.js'
-                }
-            }
-        },<% } %><% if (coffee) { %>
-
-        // Compiles CoffeeScript to JavaScript
-        coffee: {
-            dist: {
-                files: [{
-                    expand: true,
-                    cwd: '<%%= config.app %>/scripts',
-                    src: '{,*/}*.{coffee,litcoffee,coffee.md}',
-                    dest: '.tmp/scripts',
-                    ext: '.js'
-                }]
-            },
-            test: {
-                files: [{
-                    expand: true,
-                    cwd: 'test/spec',
-                    src: '{,*/}*.{coffee,litcoffee,coffee.md}',
-                    dest: '.tmp/spec',
-                    ext: '.js'
-                }]
-            }
-        },<% } %><% if (includeSass) { %>
-
-        // Compiles Sass to CSS and generates necessary files if requested
-        sass: {
-            options: {<% if (includeLibSass) { %>
-                includePaths: [
-                    'bower_components'
-                ]<% } if (includeRubySass) { %>
-                loadPath: [
-                    'bower_components'
-                ]<% } %>
-            },
-            dist: {
-                files: [{
-                    expand: true,
-                    cwd: '<%%= config.app %>/styles',
-                    src: ['*.scss'],
-                    dest: '.tmp/styles',
-                    ext: '.css'
-                }]
-            },
-            server: {
-                files: [{
-                    expand: true,
-                    cwd: '<%%= config.app %>/styles',
-                    src: ['*.scss'],
-                    dest: '.tmp/styles',
-                    ext: '.css'
-                }]
-            }
-        },<% } %>
-
-        // Add vendor prefixed styles
-        autoprefixer: {
-            options: {
-                browsers: ['last 1 version']
-            },
-            dist: {
-                files: [{
-                    expand: true,
-                    cwd: '.tmp/styles/',
-                    src: '{,*/}*.css',
-                    dest: '.tmp/styles/'
-                }]
-            }
-        },
-
-        // Automatically inject Bower components into the HTML file
-        bowerInstall: {
-            app: {
-                src: ['<%%= config.app %>/index.html'],
-                ignorePath: '<%%= config.app %>/',<% if (includeSass) { %>
-                exclude: ['bower_components/bootstrap-sass-official/vendor/assets/javascripts/bootstrap.js']<% } else { %>
-                exclude: ['bower_components/bootstrap/dist/js/bootstrap.js']<% } %>
-            }<% if (includeSass) { %>,
-            sass: {
-                src: ['<%%= config.app %>/styles/{,*/}*.{scss,sass}'],
-                ignorePath: '<%%= config.app %>/bower_components/'
-            }<% } %>
-        },
-
-        // Renames files for browser caching purposes
-        rev: {
-            dist: {
-                files: {
-                    src: [
-                        '<%%= config.dist %>/scripts/{,*/}*.js',
-                        '<%%= config.dist %>/styles/{,*/}*.css',
-                        '<%%= config.dist %>/images/{,*/}*.*',
-                        '<%%= config.dist %>/styles/fonts/{,*/}*.*',
-                        '<%%= config.dist %>/*.{ico,png}'
-                    ]
-                }
-            }
-        },
-
-        // Reads HTML for usemin blocks to enable smart builds that automatically
-        // concat, minify and revision files. Creates configurations in memory so
-        // additional tasks can operate on them
-        useminPrepare: {
-            options: {
-                dest: '<%%= config.dist %>'
-            },
-            html: '<%%= config.app %>/index.html'
-        },
-
-        // Performs rewrites based on rev and the useminPrepare configuration
-        usemin: {
-            options: {
-                assetsDirs: ['<%%= config.dist %>', '<%%= config.dist %>/images']
-            },
-            html: ['<%%= config.dist %>/{,*/}*.html'],
-            css: ['<%%= config.dist %>/styles/{,*/}*.css']
-        },
-
-        // The following *-min tasks produce minified files in the dist folder
-        imagemin: {
-            dist: {
-                files: [{
-                    expand: true,
-                    cwd: '<%%= config.app %>/images',
-                    src: '{,*/}*.{gif,jpeg,jpg,png}',
-                    dest: '<%%= config.dist %>/images'
-                }]
-            }
-        },
-
-        svgmin: {
-            dist: {
-                files: [{
-                    expand: true,
-                    cwd: '<%%= config.app %>/images',
-                    src: '{,*/}*.svg',
-                    dest: '<%%= config.dist %>/images'
-                }]
-            }
-        },
-
-        htmlmin: {
-            dist: {
-                options: {
-                    collapseBooleanAttributes: true,
-                    collapseWhitespace: true,
-                    removeAttributeQuotes: true,
-                    removeCommentsFromCDATA: true,
-                    removeEmptyAttributes: true,
-                    removeOptionalTags: true,
-                    removeRedundantAttributes: true,
-                    useShortDoctype: true
-                },
-                files: [{
-                    expand: true,
-                    cwd: '<%%= config.dist %>',
-                    src: '{,*/}*.html',
-                    dest: '<%%= config.dist %>'
-                }]
-            }
-        },
-
-        // By default, your `index.html`'s <!-- Usemin block --> will take care of
-        // minification. These next options are pre-configured if you do not wish
-        // to use the Usemin blocks.
-        // cssmin: {
-        //     dist: {
-        //         files: {
-        //             '<%%= config.dist %>/styles/main.css': [
-        //                 '.tmp/styles/{,*/}*.css',
-        //                 '<%%= config.app %>/styles/{,*/}*.css'
-        //             ]
-        //         }
-        //     }
-        // },
-        // uglify: {
-        //     dist: {
-        //         files: {
-        //             '<%%= config.dist %>/scripts/scripts.js': [
-        //                 '<%%= config.dist %>/scripts/scripts.js'
-        //             ]
-        //         }
-        //     }
-        // },
-        // concat: {
-        //     dist: {}
-        // },
-
-        // Copies remaining files to places other tasks can use
-        copy: {
-            dist: {
-                files: [{
-                    expand: true,
-                    dot: true,
-                    cwd: '<%%= config.app %>',
-                    dest: '<%%= config.dist %>',
-                    src: [
-                        '*.{ico,png,txt}',
-                        '.htaccess',
-                        'images/{,*/}*.webp',
-                        '{,*/}*.html',
-                        'styles/fonts/{,*/}*.*'
-                    ]
-                }<% if (includeBootstrap) { %>, {
-                    expand: true,
-                    dot: true,<% if (includeSass) { %>
-                    cwd: '.',
-                    src: ['bower_components/bootstrap-sass-official/vendor/assets/fonts/bootstrap/*.*'],<% } else { %>
-                    cwd: 'bower_components/bootstrap/dist',
-                    src: ['fonts/*.*'],<% } %>
-                    dest: '<%%= config.dist %>'
-                }<% } %>]
-            },
-            styles: {
-                expand: true,
-                dot: true,
-                cwd: '<%%= config.app %>/styles',
-                dest: '.tmp/styles/',
-                src: '{,*/}*.css'
-            }
-        },<% if (includeModernizr) { %>
-
-        // Generates a custom Modernizr build that includes only the tests you
-        // reference in your app
-        modernizr: {
-            devFile: 'bower_components/modernizr/modernizr.js',
-            outputFile: '<%%= config.dist %>/scripts/vendor/modernizr.js',
-            files: [
-                '<%%= config.dist %>/scripts/{,*/}*.js',
-                '<%%= config.dist %>/styles/{,*/}*.css',
-                '!<%%= config.dist %>/scripts/vendor/*'
-            ],
-            uglify: true
-        },<% } %>
-
-        // Run some tasks in parallel to speed up build process
-        concurrent: {
-            server: [<% if (includeSass) { %>
-                'sass:server',<% } if (coffee) {  %>
-                'coffee:dist',<% } %>
-                'copy:styles'
-            ],
-            test: [<% if (coffee) { %>
-                'coffee',<% } %>
-                'copy:styles'
-            ],
-            dist: [<% if (coffee) { %>
-                'coffee',<% } if (includeSass) { %>
-                'sass',<% } %>
-                'copy:styles',
-                'imagemin',
-                'svgmin'
-            ]
-        }
+        config : config,
+        watch : <%= configs.watch %>,
+        connect : <%= configs.connect %>,
+        clean : <%= configs.clean %>,
+        jshint : <%= configs.jshint %>,<% if (testFramework === 'mocha') { %>
+        mocha : <%= configs.mocha %>,<% } else if (testFramework === 'jasmine') { %>
+        jasmine : <%= configs.jasmine %>,<% } %><% if (coffee) { %>
+        coffee : <%= configs.coffee %>,<% } %><% if (includeSass) { %>
+        sass : <%= configs.sass %>,<% } %>
+        autoprefixer : <%= configs.autoprefixer %>,
+        bowerInstall : <%= configs.bowerInstall %>,
+        rev : <%= configs.rev %>,
+        useminPrepare : <%= configs.useminPrepare %>,
+        usemin : <%= configs.usemin %>,
+        imagemin : <%= configs.imagemin %>,
+        svgmin: <%= configs.svgmin %>,
+        htmlmin: <%= configs.htmlmin %>,
+        copy: <%= configs.copy %>,<% if (includeModernizr) { %>
+        modernizr: <%= configs.modernizr %>,<% } %>
+        concurrent: <%= configs.concurrent %>
     });
-
-
+    <% } %>
     grunt.registerTask('serve', function (target) {
         if (target === 'dist') {
             return grunt.task.run(['build', 'connect:dist:keepalive']);
