+++ app/templates/_module.js
@@ -1,0 +1,3 @@
+module.exports = function(grunt, config){
+  return <%= taskConfig %>;
+};
