+++ app/templates/_package.json
@@ -29,7 +29,8 @@
     "grunt-concurrent": "~0.5.0",
     "load-grunt-tasks": "~0.4.0",
     "time-grunt": "~0.3.1",
-    "jshint-stylish": "~0.1.5"
+    "jshint-stylish": "~0.1.5"<% if (splitConfigs) {  %>,
+    "load-grunt-configs" : "~0.3.3"<% } %>
   },
   "engines": {
     "node": ">=0.10.0"
