+++ app/templates/config/_autoprefixer.js
@@ -1,0 +1,14 @@
+{
+            // Add vendor prefixed styles
+            options: {
+                browsers: ['last 1 version']
+            },
+            dist: {
+                files: [{
+                    expand: true,
+                    cwd: '.tmp/styles/',
+                    src: '{,*/}*.css',
+                    dest: '.tmp/styles/'
+                }]
+            }
+        }
