+++ app/templates/config/_clean.js
@@ -1,0 +1,14 @@
+{
+            // Empties folders to start fresh
+            dist: {
+                files: [{
+                    dot: true,
+                    src: [
+                        '.tmp',
+                        '<%%= config.dist %>/*',
+                        '!<%%= config.dist %>/.git*'
+                    ]
+                }]
+            },
+            server: '.tmp'
+        }
