+++ app/templates/config/_config.js
@@ -1,0 +1,4 @@
+{
+            app: config.app,
+            dist: config.dist
+        }
