+++ app/templates/config/_htmlmin.js
@@ -1,0 +1,20 @@
+{
+            dist: {
+                options: {
+                    collapseBooleanAttributes: true,
+                    collapseWhitespace: true,
+                    removeAttributeQuotes: true,
+                    removeCommentsFromCDATA: true,
+                    removeEmptyAttributes: true,
+                    removeOptionalTags: true,
+                    removeRedundantAttributes: true,
+                    useShortDoctype: true
+                },
+                files: [{
+                    expand: true,
+                    cwd: '<%%= config.dist %>',
+                    src: '{,*/}*.html',
+                    dest: '<%%= config.dist %>'
+                }]
+            }
+        }
