+++ app/templates/config/_jasmine.js
@@ -1,0 +1,8 @@
+{
+            // Jasmine testing framework configuration options
+            all: {
+                options: {
+                    specs: 'test/spec/{,*/}*.js'
+                }
+            }
+        }
