+++ app/templates/config/_mocha.js
@@ -1,0 +1,9 @@
+{
+            // Mocha testing framework configuration options
+            all: {
+                options: {
+                    run: true,
+                    urls: ['http://<%%= connect.test.options.hostname %>:<%%= connect.test.options.port %>/index.html']
+                }
+            }
+        }
