+++ app/templates/config/_svgmin.js
@@ -1,0 +1,10 @@
+{
+            dist: {
+                files: [{
+                    expand: true,
+                    cwd: '<%%= config.app %>/images',
+                    src: '{,*/}*.svg',
+                    dest: '<%%= config.dist %>/images'
+                }]
+            }
+        }
