+++ app/templates/config/_useminPrepare.js
@@ -1,0 +1,9 @@
+{
+            // Reads HTML for usemin blocks to enable smart builds that automatically
+            // concat, minify and revision files. Creates configurations in memory so
+            // additional tasks can operate on them
+            options: {
+                dest: '<%%= config.dist %>'
+            },
+            html: '<%%= config.app %>/index.html'
+        }
