+++ test/test.js
@@ -97,4 +97,46 @@ describe('Webapp generator test', function () {
       done();
     });
   });
+
+  it('creates expected files when --split-configs flag set', function (done) {
+    var expected = [
+      ['bower.json', /"name": "temp"/],
+      ['package.json', /"name": "temp"/],
+      'Gruntfile.js',
+      'app/404.html',
+      'app/favicon.ico',
+      'app/robots.txt',
+      'app/index.html',
+      'app/scripts/main.js',
+      'app/styles/main.scss',
+      'config/autoprefixer.js',
+      'config/bowerInstall.js',
+      'config/clean.js',
+      'config/sass.js',
+      'config/concurrent.js',
+      'config/connect.js',
+      'config/copy.js',
+      'config/htmlmin.js',
+      'config/imagemin.js',
+      'config/jshint.js',
+      'config/rev.js',
+      'config/svgmin.js',
+      'config/usemin.js',
+      'config/useminPrepare.js',
+      'config/watch.js',
+      'config/config.js'
+    ];
+
+    helpers.mockPrompt(this.webapp, {
+      features: ['includeSass' ]
+    });
+
+    this.webapp.coffee = false;
+    this.webapp.splitConfigs = true;
+
+    this.webapp.run({}, function () {
+      helpers.assertFiles(expected);
+      done();
+    });
+  });
 });
